requirements.txt
anyio==3.6.2
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.12.7
click==8.1.3
cmake==3.26.3
dlib==19.24.1
face-recognition==1.3.0
face-recognition-models==0.3.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
numpy==1.24.2
Pillow==9.5.0
python-telegram-bot==13.14
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tornado==6.1
tzdata==2023.3
tzlocal==4.3
